impure: Truncate current working path if too long
Previously the path would be just restricted to the last three
directories, but if the name of these happen to be long the prompt could
still overflow. This is exacerbated if git information is also being
displayed. Now trailing path components are trimmed further if it would
result in a prompt line that is longer than the terminal is wide.