zsh: Change command to rename tmux windows
Some programs set the window name using a terminal escape code, which
was interpreted as a name set by the user. As such this name would
remain until the user manually reset. If the user had actually set a
name it would override the name set by the user.

To fix this the script now uses `tmux rename-window` instead of the
escape sequence when renaming windows.