zsh: Fix config loading between vim and nvim
VIMINIT is used so that vim can respect XDG directory configurations,
however this variable is also read by nvim if is set. To fix this vim is
aliased to include VIMINIT.

Within the environment variables EDITOR is set depending upon nvim
availability. SUDO_EDITOR is always vim, as nvim is not going to be in
the $path of the root user. Besides, one should not need elevated editor
privileges very often or for very long, so I can live without any
special config.